+++ src/types.ts
@@ -1,7 +1,7 @@
 import type { ConnectOptions } from './connect.js'
 import type { ExtraTableConfigIndexBase, PrimaryIndex, QueryForIndex } from './indexes.js'
 import type { SyncOptions, SyncReturnType } from './sync.js'
-import type { DeepPartial, DeepValue, DeepValuePartial, DistOmit, Empty, PushEmpty } from './util.js'
+import type { DeepPartial, DeepValue, DeepValuePartial, DistOmit, Empty, PushEmpty, RMinMaxVal } from './util.js'
 import type { FieldSelector, InsertOptions, MasterPool, R, RDatabase, RDatum, RSelection, RSingleSelection, RStream, RTable, RValue, UpdateOptions, WriteResult } from 'rethinkdb-ts'
 
 /** see: https://github.com/rethinkdb/rethinkdb/issues/2884#issuecomment-65291774 */
@@ -19,7 +19,7 @@ export type RStreamExtra<Config extends ExtraTableConfig<any, any>, T> = Omit<RS
   <A extends keyof T>(attribute: RValue<A>): RStreamExtra<Config, T[A]>,
   (n: RValue<number>): RDatum<T>,
   
-  orderBy(...fieldOrIndex: Array<FieldSelector<T> | { index: keyof Config['indexes'] }>): RStreamExtra<Config, T>
+  orderBy(...fieldOrIndex: Array<FieldSelector<T> | { index: keyof Config['indexes'] | (string & {}) }>): RStreamExtra<Config, T>
 }
 
 export type RSingleSelectionExtra<_Config extends ExtraTableConfig<any, any>, T, TN = Exclude<T, null>> = Omit<RSingleSelection<T>, 'update' | 'replace'> & {
@@ -48,7 +48,7 @@ export type RSelectionExtra<Config extends ExtraTableConfig<any, any>, T> = Omit
   replace(obj: RValue<T>, options?: UpdateOptions): RDatum<WriteResult<T>>,
   replace<Previous = T>(replacer: (previous: RDatum<Previous>) => RValue<DeepValue<T>>, options?: UpdateOptions): RDatum<WriteResult<T>>,
 
-  orderBy(...fieldOrIndex: Array<FieldSelector<T> | { index: keyof Config['indexes'] }>): RSelectionExtra<Config, T>
+  orderBy(...fieldOrIndex: Array<FieldSelector<T> | { index: keyof Config['indexes'] | (string & {}) }>): RSelectionExtra<Config, T>
 }
 
 export type RTableExtra<Config extends ExtraTableConfig<any, any>> = Omit<RTable<Config['type']>, 'get' | 'getAll' | 'between' | 'insert' | 'update' | 'replace' | 'orderBy'> & {
@@ -57,12 +57,20 @@ export type RTableExtra<Config extends ExtraTableConfig<any, any>> = Omit<RTable
   
   get(key: Config['type']['id'] | RDatum<Config['type']['id']>): RSingleSelectionExtra<Config, Config['type'] | null>,
 
-  getAll<I extends keyof Config['indexes']>(key: QueryForIndex<Config, I>, options: { index: I }): RSelectionExtra<Config, Config['type']>,
-  getAll<I extends (keyof Config['indexes'] | PrimaryIndex) = PrimaryIndex>(...keys: Array<QueryForIndex<Config, I> | { index: I }>): RSelectionExtra<Config, Config['type']>,
+  getAll<I extends (keyof Config['indexes'] | PrimaryIndex) = PrimaryIndex>(
+    ...args: [
+      QueryForIndex<Config, I>,
+      ...QueryForIndex<Config, I>[]
+    ] | [
+      QueryForIndex<Config, I>,
+      ...QueryForIndex<Config, I>[],
+      { index: I }
+    ]
+  ): RSelectionExtra<Config, Config['type']>,
 
   between<I extends (keyof Config['indexes'] | PrimaryIndex) = PrimaryIndex>(
-    lowKey: QueryForIndex<Config, I>,
-    highKey: QueryForIndex<Config, I>,
+    lowKey: QueryForIndex<Config, I, RMinMaxVal>,
+    highKey: QueryForIndex<Config, I, RMinMaxVal>,
     options?: { index?: I, leftBound?: 'open' | 'closed', rightBound?: 'open' | 'closed' }
   ): RSelectionExtra<Config, Config['type']>,
 
@@ -82,13 +90,13 @@ export type RTableExtra<Config extends ExtraTableConfig<any, any>> = Omit<RTable
   replace(obj: RValue<Config['type']>, options?: UpdateOptions): RDatum<WriteResult<Config['type']>>,
   replace<Previous = Config['type']>(replacer: (previous: RDatum<Previous>) => RValue<DeepValue<Config['type']>>, options?: UpdateOptions): RDatum<WriteResult<Config['type']>>,
 
-  orderBy(...fieldOrIndex: Array<FieldSelector<Config['type']> | { index: keyof Config['indexes'] }>): RTableExtra<Config>
+  orderBy(...fieldOrIndex: Array<FieldSelector<Config['type']> | { index: keyof Config['indexes'] | (string & {}) }>): RTableExtra<Config>
 }
 
 export type RDatabaseExtraConfigs = Record<string, Record<string, ExtraTableConfig<any, any>>>
 
 export type RDatabaseExtra<Configs extends RDatabaseExtraConfigs, Database extends keyof Configs> = Omit<RDatabase, 'db' | 'table'> & {
-  table<T extends keyof Configs[Database]>(tableName: T): RTableExtra<Configs[Database][T]>
+  table<T extends keyof Configs[Database] | (string & {})>(tableName: T): T extends keyof Configs[Database] ? RTableExtra<Configs[Database][T]> : RTable
 }
 
 export type RExtra<Configs extends RDatabaseExtraConfigs, DefaultDB extends string> = Omit<R, 'asc' | 'desc' | 'literal' | 'db' | 'table'> & {
@@ -97,9 +105,9 @@ export type RExtra<Configs extends RDatabaseExtraConfigs, DefaultDB extends stri
   literal(): Empty,
   literal<T>(obj: T): RDatum<T>,
 
-  db<D extends keyof Configs>(dbName: D): RDatabaseExtra<Configs, D>,
-  db(dbName: string): RDatabase,
-  table<T extends (DefaultDB extends keyof Configs ? keyof Configs[DefaultDB] : string)>(tableName: T): DefaultDB extends keyof Configs ? RTableExtra<Configs[DefaultDB][T]> : RTable,
+  db<D extends keyof Configs | (string & {})>(dbName: D): D extends keyof Configs ? RDatabaseExtra<Configs, D> : RDatabase,
+  table<T extends (DefaultDB extends keyof Configs ? keyof Configs[DefaultDB] : never) | (string & {})>(tableName: T):
+    DefaultDB extends keyof Configs ? T extends keyof Configs[DefaultDB] ? RTableExtra<Configs[DefaultDB][T]> : RTable : RTable,
 
   extra: {
     connect(options?: ConnectOptions<DefaultDB>): Promise<MasterPool>,